100x Cash by Akash Das
This is a game in which you need to collect Rs.10 Coins. But if you fail to collect the coin it hits the flower, which will deduct Rs.50 from your account.  To Play you just need to use the right and left arrows from the keyboard, to move the piggybank.  Just play! and have fun.  Developed by Akash Das